Split out the OASys risk summary view component into it's own class. This will be used by both the edit and view OASys risk information pages.
